backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
pydantic-settings==2.6.1
cloudinary==1.41.0
faster-whisper==1.1.1
pydub==0.25.1
langchain-groq==0.2.0
slowapi==0.1.9
httpx==0.28.1
pytest==8.3.4
pytest-mock==3.14.0
groq<1.0.0,>=0.4.1